ignore errors deleting the analysis driver cache.
BUG=
R=brianwilkerson@google.com, scheglov@google.com

Review-Url: https://codereview.chromium.org/2992203002 .

+++ pkg/analysis_server/benchmark/benchmarks.dart
@@ -225,5 +225,11 @@ void deleteServerCache() {
   // ~/.dartServer/.analysis-driver/
   ResourceProvider resourceProvider = PhysicalResourceProvider.INSTANCE;
   Folder stateLocation = resourceProvider.getStateLocation('.analysis-driver');
-  stateLocation.delete();
+  try {
+    if (stateLocation.exists) {
+      stateLocation.delete();
+    }
+  } catch (e) {
+    // ignore any exception
+  }
 }
